+++ src/main/java/ru/mail/polis/lsm/artemdrozdov/LsmDAO.java
@@ -66,12 +66,12 @@ public class LsmDAO implements DAO {
 
     @Override
     public Iterator<Record> range(@Nullable ByteBuffer fromKey, @Nullable ByteBuffer toKey) {
-        return filterTombstones(find(fromKey, toKey), true);
+        return TombstoneFilter.filterTombstones(find(fromKey, toKey), true);
     }
 
     @Override
     public Iterator<Record> rangeWithTombstone(@Nullable ByteBuffer fromKey, @Nullable ByteBuffer toKey) {
-        return filterTombstones(find(fromKey, toKey), false);
+        return TombstoneFilter.filterTombstones(find(fromKey, toKey), false);
     }
 
     /**
@@ -229,35 +229,4 @@ public class LsmDAO implements DAO {
         return new MergeIterator(left, right);
     }
 
-    private static Iterator<Record> filterTombstones(Iterator<Record> iterator, final boolean noTombstone) {
-        PeekingIterator delegate = new PeekingIterator(iterator);
-        return new Iterator<>() {
-            @Override
-            public boolean hasNext() {
-                for (;;) {
-                    Record peek = delegate.peek();
-                    if (peek == null) {
-                        return false;
-                    }
-                    if (noTombstone) {
-                        if (!peek.isTombstone()) {
-                            return true;
-                        }
-                    } else {
-                        return true;
-                    }
-
-                    delegate.next();
-                }
-            }
-
-            @Override
-            public Record next() {
-                if (!hasNext()) {
-                    throw new NoSuchElementException("No elements");
-                }
-                return delegate.next();
-            }
-        };
-    }
 }
